refactor(calendars): move CalendarForm to its own shared component
